requirements.txt
Flask==2.1.2
Flask_JSONLocale==1.11
Flask_SocketIO==5.1.2
MarkupSafe==2.1.1
PyYAML==6.0
SQLAlchemy==1.4.36
